paul/adds viewmodel
